requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
Flask_WTF==1.1.1
WTForms==3.0.1
python-dotenv~=1.0.0